requirements.txt
easyocr==1.7.1
numpy==1.24.4
pillow==10.1.0

